requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
multidict==5.2.0
numpy==1.21.4
numpy-stl==2.16.3
oauthlib==3.1.1
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.6
python-utils==2.5.6
pytorch-lightning==1.5.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
stl==0.0.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
torch==1.10.0
torchmetrics==0.6.0
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0

matplotlib~=3.5.0
moviepy~=1.0.3
torchvision~=0.11.1
selenium~=4.1.0
python-dotenv~=0.19.2
omegaconf~=2.1.1